i think i'm about done with modernizing and organizing this file
next task is to write a complete test suite for selection, so i can
hopefully just run tests against it to get it fixed properly, instead of
having to manually debug it every time i make a change.

and then i'm going to get to trying to reduce this down to as small of a
component as possible, once the tests are complete.